Refactor Google api client/authorization
Instead of having an intialiser setting a global authorization for all
the Google Api Client requests, we extract the Api Client authorization
into its own singleton class.

That class will get the correct authorization and refresh the token when
its expired.

The different Services and Jobs relying in this Google client will call
this singleton instance to authorise their requests.

The intention of this is to stop setting global values at Rails
initialization level, as it may have unintended consequences between
Google Analytics internals mixing with the global authorization set for
Google Drive and Search Index APIs.

+++ app/services/publishers/document_virus_check.rb
@@ -7,28 +7,37 @@ class Publishers::DocumentVirusCheck
   FILE_VIRUS_STATUS_CODE = 403 # 403 Permission denied when acknowledge_abuse: false
 
   def initialize(file)
+    @api_client = GoogleApiClient.instance
+    return if api_client.missing_key?
+
+    drive_service.authorization = api_client.authorization
     @file = file
   end
 
   def safe?
+    return false if api_client.missing_key?
+
     drive_service.get_file(
       uploaded_file.id,
       acknowledge_abuse: false,
       download_dest: uploaded_file.id.to_s,
     )
+
     true
   rescue Google::Apis::ClientError => e
     return false if e.status_code == FILE_VIRUS_STATUS_CODE
 
     raise e
   ensure
-    FileUtils.rm_rf(uploaded_file.id.to_s)
-    drive_service.delete_file(uploaded_file.id)
+    unless api_client.missing_key?
+      FileUtils.rm_rf(uploaded_file.id.to_s)
+      drive_service.delete_file(uploaded_file.id)
+    end
   end
 
   private
 
-  attr_reader :file
+  attr_reader :file, :api_client
 
   def uploaded_file
     @uploaded_file ||= drive_service.create_file(
